Configurei a parte de Exceções do programa

+++ src/main/java/com/example/curso_springboot/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.example.curso_springboot.service;
 
+import com.example.curso_springboot.exceptions.RecursoNaoEncontradoException;
 import com.example.curso_springboot.model.Produto;
 import com.example.curso_springboot.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -21,8 +22,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID " + id + " não encontrado!"));
     }
 
     public Produto salvarProduto(Produto produto){
@@ -30,6 +32,10 @@ public class ProdutoService {
     }
 
     public void deletarProduto(long id){
+
+        if(!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto com ID " + id + " não encontrado!");
+        }
         produtoRepository.deleteAllById(Collections.singleton(id));
     }
 }
